Extracted QuantityResponse Signed-off-by: Richard Lavoie <lavoie.richard@gmail.com>
