+++ addons/binding/org.openhab.binding.lgtvserial/src/main/java/org/openhab/binding/lgtvserial/internal/protocol/serial/commands/TemperatureValueCommand.java
@@ -14,8 +14,8 @@ package org.openhab.binding.lgtvserial.internal.protocol.serial.commands;
 
 import org.eclipse.smarthome.core.library.types.QuantityType;
 import org.eclipse.smarthome.core.library.unit.SIUnits;
-import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.lgtvserial.internal.protocol.serial.LGSerialResponse;
+import org.openhab.binding.lgtvserial.internal.protocol.serial.responses.QuantityResponse;
 
 /**
  * This class handles the temperature value D/N command.
@@ -34,34 +34,4 @@ public class TemperatureValueCommand extends BaseDecimalCommand {
         return new QuantityResponse(set, success, new QuantityType<>(Integer.parseInt(data, 16), SIUnits.CELSIUS));
     }
 
-    public static class QuantityResponse implements LGSerialResponse {
-
-        private int setId;
-
-        private boolean success;
-
-        private State state;
-
-        public QuantityResponse(int setId, boolean success, QuantityType<?> state) {
-            this.setId = setId;
-            this.success = success;
-            this.state = state;
-        }
-
-        @Override
-        public int getSetID() {
-            return setId;
-        }
-
-        @Override
-        public State getState() {
-            return state;
-        }
-
-        @Override
-        public boolean isSuccess() {
-            return success;
-        }
-
-    }
 }
